LaserFields

Documentation for LaserFields.

LaserFields.envelope_fourier

LaserFields.@select_param

LaserFields.envelope_fourier — Function

Return the fourier transform of the envelope of the laser field. We write the whole pulse as f(t) = (env(t) exp(i(phi0 + ω0tr + chirp*tr^2)) + c.c. ) / 2im, where tr = t-tpeak For the fourier transform of the envelope, we include the chirp term exp(i chirp (t-tpeak)^2) in the envelope, so that its fourier transform is a complex function. However, for unchirped pulses, the result will be purely real.

LaserFields.@select_param — Macro

Get a value from one of several possibly defined parameters, or a default value if none are defined. Makes sure that at most one of the options are specified, and that at least one is given if there is no default value.

sample use: `x = @select_param args (x => args.x, x_squared => sqrt(x_squared), 0.)`

This makes it possible to call a function with either x, x_squared, or neither, and in the last case, x is set to 0.