LaserFields

Documentation for LaserFields.

LaserFields.envelope_fourier

LaserFields.envelope_fourier — Function

return the fourier transform of the envelope of the laser field. we write the whole pulse as f(t) = (env(t) exp(i(phi0 + ω0tp + chirp*tp^2)) + c.c. ) / 2im, where tp = t-tpeak for the fourier transform of the envelope, we include the chirp term exp(i chirp (t-tpeak)^2) in the envelope, so that its fourier transform is a complex function. however, for unchirped pulses, the result will be purely real!

for the various calculations, see chirped_fourier.nb in the mathematica directory.